+++ sdk/communication/communication-network-traversal/recordings/node/communicationnetworkingclient_playbacklive_aad/recording_successfully_gets_a_turn_credential.js
@@ -1,0 +1,94 @@
+let nock = require('nock');
+
+module.exports.hash = "79bce32afb53073cd137b3551cf7e47a";
+
+module.exports.testInfo = {"uniqueName":{},"newDate":{}}
+
+nock('https://endpoint', {"encodedQueryParams":true})
+  .post('/identities')
+  .query(true)
+  .reply(201, {"identity":{"id":"sanitized"}}, [
+  'Transfer-Encoding',
+  'chunked',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'Request-Context',
+  'appId=',
+  'MS-CV',
+  '2kzqf1TPSUK6hHSXhJfSwQ.0',
+  'Strict-Transport-Security',
+  'max-age=2592000',
+  'x-ms-client-request-id',
+  'sanitized',
+  'api-supported-versions',
+  '2020-07-20-preview2, 2021-02-22-preview1, 2021-03-07, 2021-03-31-preview1',
+  'X-Processing-Time',
+  '59ms',
+  'X-Cache',
+  'CONFIG_NOCACHE',
+  'X-Azure-Ref',
+  '0r1ykYAAAAADGbLEOvDQRTJ3L2iqY3sFXV1NURURHRTA4MTMAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx',
+  'Date',
+  'Wed, 19 May 2021 00:32:47 GMT'
+]);
+
+nock('https://endpoint', {"encodedQueryParams":true})
+  .post('/SomeTenantId/oauth2/v2.0/token', "response_type=token&grant_type=client_credentials&client_id=SomeClientId&client_secret=SomeClientSecret&scope=https%3A%2F%2Fcommunication.azure.com%2F%2F.default")
+  .reply(200, {"token_type":"Bearer","expires_in":3599,"ext_expires_in":3599,"access_token":"sanitized"}, [
+  'Cache-Control',
+  'no-store, no-cache',
+  'Pragma',
+  'no-cache',
+  'Content-Length',
+  '1325',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'Expires',
+  '-1',
+  'Strict-Transport-Security',
+  'max-age=31536000; includeSubDomains',
+  'X-Content-Type-Options',
+  'nosniff',
+  'P3P',
+  'CP="DSP CUR OTPi IND OTRi ONL FIN"',
+  'x-ms-request-id',
+  'sanitized',
+  'x-ms-ests-server',
+  '2.1.11722.21 - WUS2 ProdSlices',
+  'Set-Cookie',
+  'fpc=AtC7ogwuoqxIpereXjWWPolWvwPgAQAAAK9TNtgOAAAA; expires=Fri, 18-Jun-2021 00:32:48 GMT; path=/; secure; HttpOnly; SameSite=None',
+  'Set-Cookie',
+  'x-ms-gateway-slice=estsfd; path=/; secure; samesite=none; httponly',
+  'Set-Cookie',
+  'stsservicecookie=estsfd; path=/; secure; samesite=none; httponly',
+  'Date',
+  'Wed, 19 May 2021 00:32:48 GMT'
+]);
+
+nock('https://endpoint', {"encodedQueryParams":true})
+  .post('/turn/sanitized/:issueCredentials')
+  .query(true)
+  .reply(200, {"expiresOn":"2022-05-18T12:00:00.00+00:00","turnServers":[{"urls":["turn.skype.com"]},"username":"sanitized_username","credential":"sanitized_credential"}]}, [
+  'Transfer-Encoding',
+  'chunked',
+  'Content-Type',
+  'application/json; charset=utf-8',
+  'Request-Context',
+  'appId=',
+  'MS-CV',
+  '8sEAhrI/D0ObneKCEDOJtQ.0',
+  'Strict-Transport-Security',
+  'max-age=2592000',
+  'x-ms-client-request-id',
+  'sanitized',
+  'api-supported-versions',
+  '2021-02-22-preview1',
+  'X-Processing-Time',
+  '319ms',
+  'X-Cache',
+  'CONFIG_NOCACHE',
+  'X-Azure-Ref',
+  '0sFykYAAAAAAuSnYwqajEQI8uyMShKQuBV1NURURHRTA4MDYAOWZjN2I1MTktYThjYy00Zjg5LTkzNWUtYzkxNDhhZTA5ZTgx',
+  'Date',
+  'Wed, 19 May 2021 00:32:48 GMT'
+]);