+++ sdk/communication/communication-network-traversal/test/public/utils/recordedClient.ts
@@ -33,8 +33,7 @@ export interface RecordedClient<T> {
 }
 
 const replaceableVariables: { [k: string]: string } = {
-  COMMUNICATION_CONNECTION_STRING: "endpoint=https://endpoint/;accesskey=banana",
-  INCLUDE_PHONENUMBER_LIVE_TESTS: "false",
+  COMMUNICATION_LIVETEST_DYNAMIC_CONNECTION_STRING: "endpoint=https://endpoint/;accesskey=banana",
   COMMUNICATION_ENDPOINT: "https://endpoint/",
   AZURE_CLIENT_ID: "SomeClientId",
   AZURE_CLIENT_SECRET: "SomeClientSecret",
@@ -48,6 +47,14 @@ export const environmentSetup: RecorderEnvironmentSetup = {
       recording.replace(/"token"\s?:\s?"[^"]*"/g, `"token":"sanitized"`),
     (recording: string): string =>
       recording.replace(/"access_token"\s?:\s?"[^"]*"/g, `"access_token":"sanitized"`),
+    (recording: string): string =>
+      recording.replace(/"urls"\s?:\s?\[.*?\]/g, `"urls":["turn.skype.com"]}`),
+    (recording: string): string =>
+      recording.replace(/"username"\s?:\s?"[^"]*"/g, `"username":"sanitized_username"`),
+    (recording: string): string =>
+      recording.replace(/"credential"\s?:\s?"[^"]*"/g, `"credential":"sanitized_credential"`),
+    (recording: string): string =>
+      recording.replace(/"expiresOn"\s?:\s?"[^"]*"/g, `"expiresOn":"2022-05-18T12:00:00.00+00:00"`),
     (recording: string): string => recording.replace(/(https:\/\/)([^/',]*)/, "$1endpoint"),
     (recording: string): string => recording.replace(/"id"\s?:\s?"[^"]*"/g, `"id":"sanitized"`),
     (recording: string): string => {
@@ -56,8 +63,7 @@ export const environmentSetup: RecorderEnvironmentSetup = {
         "$1sanitized$2"
       );
     },
-    (recording: string): string =>
-      recording.replace(/\/identities\/[^/'",]*/, "/identities/sanitized"),
+    (recording: string): string => recording.replace(/\/turn\/[^/'",]*/, "/turn/sanitized"),
     (recording: string): string => recording.replace(/\+\d{1}\d{3}\d{3}\d{4}/g, "+18005551234"),
     (recording: string): string =>
       recording.replace(/[0-9a-f]{8}-([0-9a-f]{4}-){3}[0-9a-f]{12}/gi, "sanitized")
@@ -72,7 +78,7 @@ export function createRecordedCommunicationRelayClient(
 
   // casting is a workaround to enable min-max testing
   return {
-    client: new CommunicationRelayClient(env.COMMUNICATION_CONNECTION_STRING, {
+    client: new CommunicationRelayClient(env.COMMUNICATION_LIVETEST_DYNAMIC_CONNECTION_STRING, {
       httpClient: createTestHttpClient()
     } as CommunicationRelayClientOptions),
     recorder
@@ -84,7 +90,8 @@ export function createRecordedCommunicationRelayClientWithToken(
 ): RecordedClient<CommunicationRelayClient> {
   const recorder = record(context, environmentSetup);
   let credential: TokenCredential;
-  const endpoint = parseConnectionString(env.COMMUNICATION_CONNECTION_STRING).endpoint;
+  const endpoint = parseConnectionString(env.COMMUNICATION_LIVETEST_DYNAMIC_CONNECTION_STRING)
+    .endpoint;
   if (isPlaybackMode()) {
     credential = {
       getToken: async (_scopes) => {
@@ -138,5 +145,5 @@ function createTestHttpClient(): HttpClient {
     return requestResponse;
   };
 
-    return customHttpClient;
-  }
+  return customHttpClient;
+}