+++ sdk/communication/communication-network-traversal/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 1.0.0-beta.1 (2021-02-22-preview1)
 
+The first preview of the Azure Communication Relay Client has the following features:
+
+- get a Relay Configuration by creating a CommunicationRelayClient
+
+
 ### Added
 
 - Added CommunicationRelayClient in preview.